proposals/web_scraping_xplatform_alert_telegram/requirements.txt
python-telegram-bot==21.10
selenium==3.141.0
urllib3==1.26.16
webdriver-manager==4.0.2
pandas
